Use the manifest path for disambiguation
