+++ cmd/run.go
@@ -15,14 +15,13 @@ import (
 //Run executes a custom command on the CND container
 func Run() *cobra.Command {
 	var devPath string
-	var namespace string
 	cmd := &cobra.Command{
 		Use:   "run SCRIPT ARGS",
 		Short: fmt.Sprintf("Run a script defined in your %s file directly in your cloud native environment", config.CNDManifestFileName()),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			analytics.Send(analytics.EventRun, GetActionID())
 			defer analytics.Send(analytics.EventRunEnd, GetActionID())
-			return executeRun(namespace, devPath, args)
+			return executeRun(devPath, args)
 		},
 		Args: func(cmd *cobra.Command, args []string) error {
 			if len(args) < 1 || args[0] == "" {
@@ -34,18 +33,17 @@ func Run() *cobra.Command {
 	}
 
 	addDevPathFlag(cmd, &devPath, config.CNDManifestFileName())
-	addNamespaceFlag(cmd, &namespace)
 	return cmd
 }
 
-func executeRun(namespace, devPath string, args []string) error {
+func executeRun(devPath string, args []string) error {
 	dev, err := model.ReadDev(devPath)
 	if err != nil {
 		return err
 	}
 
 	if val, ok := dev.Scripts[args[0]]; ok {
-		return executeExec(namespace, devPath, parseArguments(val, args))
+		return executeExec(devPath, parseArguments(val, args))
 	}
 
 	return fmt.Errorf("%s is not defined in %s. [%s]", args[0], devPath, strings.Join(getScripts(dev), ", "))
