+++ cmd/up.go
@@ -71,6 +71,8 @@ func Up() *cobra.Command {
 			}
 
 			var err error
+
+			devPath = getFullPath(devPath)
 			up.Dev, err = model.ReadDev(devPath)
 			if err != nil {
 				return err
@@ -92,7 +94,7 @@ func Up() *cobra.Command {
 				up.Context, up.Cancel = context.WithCancel(context.Background())
 				up.ErrChan = make(chan error, 1)
 
-				err = up.Execute(isRetry)
+				err = up.Execute(isRetry, devPath)
 				s.Stop()
 				if err != nil {
 					return err
@@ -192,7 +194,7 @@ func (up *UpContext) WaitUntilExit() error {
 }
 
 // Execute runs all the logic for the up command
-func (up *UpContext) Execute(isRetry bool) error {
+func (up *UpContext) Execute(isRetry bool, devPath string) error {
 
 	if !syncthing.IsInstalled() {
 		fmt.Println("Installing dependencies...")
@@ -275,7 +277,7 @@ func (up *UpContext) Execute(isRetry bool) error {
 		return err
 	}
 
-	err = storage.Insert(up.Context, up.WG, up.Namespace, up.Dev, up.Sy.GUIAddress, up.Pod.Name)
+	err = storage.Insert(up.Context, up.WG, up.Namespace, up.Dev, up.Sy.GUIAddress, up.Pod.Name, devPath)
 	if err != nil {
 		if err == storage.ErrAlreadyRunning {
 			log.Infof("failed to insert new state value for %s", up.DeploymentName)
